requirements.txt
ansicon==1.89.0
bandit==1.7.4
blessed==1.19.1
colorama==0.4.5
gitdb==4.0.9
GitPython==3.1.27
inquirer==2.10.0
jinxed==1.2.0
numpy==1.23.1
pbr==5.9.0
Pillow==9.2.0
python-editor==1.0.4
PyYAML==6.0
readchar==3.1.0
scipy==1.9.0
six==1.16.0
smmap==5.0.0
stevedore==4.0.0
wcwidth==0.2.5
